UIntGadget trait + macro implementation

+++ r1cs/gadgets/std/src/lib.rs
@@ -72,7 +72,7 @@ pub mod prelude {
     pub use crate::{
         alloc::*,
         bits::{
-            boolean::Boolean, uint32::UInt32, uint8::UInt8, FromBitsGadget, ToBitsGadget,
+            boolean::Boolean, uint32::UInt32, UInt8, FromBitsGadget, ToBitsGadget,
             ToBytesGadget,
         },
         eq::*,
